pyproject.toml
[project]
name = "beancount_cc_importers"
version = "0.1.10"
description = "Beancount importer plugin"
authors = [
    {name = "Chao Chen", email = "wenbushi@gmail.com"},
]
dependencies = [
    "lxml>=4.9.3",
    "beancount>=2.3.6",
    "python-dateutil>=2.9.0.post0",
    "azure-ai-documentintelligence==1.0.0b2",
    "azure-ai-formrecognizer>=3.3.3",
    "beangulp>=0.1.1",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
excludes = ["./**/.git"]
package-dir = "src"
includes = ["src/beancount_cc_importers"]
source-includes = ["LICENSE", "README.md"]

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.2",
]

[tool.pdm.scripts]
pytest = "pytest"

[tool.pytest.ini_options]
pythonpath = [
    "src"
]

addopts = "-s -r aR"
testpaths = [
    "tests/",
]